refactor: jobを別のクラスに切り出し
