+++ src/main/kotlin/dev/usbharu/hideout/service/ap/job/APReceiveFollowJobService.kt
@@ -1,0 +1,8 @@
+package dev.usbharu.hideout.service.ap.job
+
+import dev.usbharu.hideout.domain.model.job.ReceiveFollowJob
+import kjob.core.job.JobProps
+
+interface APReceiveFollowJobService {
+    suspend fun receiveFollowJob(props: JobProps<ReceiveFollowJob>)
+}
